un-exposed GdfBindings
